Allow reviewer to give marks in the steps of 0.5.

+++ Silicus.Ensure/Silicus.Ensure.Models/Test/TestSummaryBusinessModel.cs
@@ -11,7 +11,7 @@ namespace Silicus.Ensure.Models.Test
         public TestSummaryBasicDetails Practical { get; set; }
         public TestSummaryBasicDetails Objective { get; set; }
         public int TotalMaximumMarks { get; set; }
-        public int TotalObtainedMarks { get; set; }
+        public decimal TotalObtainedMarks { get; set; }
         public decimal Percentage { get; set; }
         public int TimeAllotted { get; set; }
         public int TimeTaken { get; set; }
